requirements.txt
nibabel>=2.1.0
scikit-learn>=0.17
nilearn>=0.3.1
pandas>=0.20
numpy>=1.9
seaborn>=0.8.0
matplotlib>=2.0
nipy>=0.4.0
pybids==0.3
nipype==0.13.1
